refactor: split 1259-line bridge_net.go into conn/listener/handshake/dialer/proxy files (SRP)
